Change findPageFile function's isWriteable check to fileExists check
I find out whether customAppFile exists or not with the findPageFile function, and the internal logic was to check if the 'pages / _app.js file is writable'.

The 'pages' folder can be 'read only'.

For example, when a Dockerfile looks like

```
FROM node: 12.16.0-alpine
...
COPY . .
next build
```

The `docker build .` command may fail even if you are building locally.

In addition, there is a possibility of failure in the CI / CD environment, so I suggest changing 'read only' to be available.